修复处理时间间隔时忘记转换类型的错误

+++ do.py
@@ -27,8 +27,8 @@ while True:
         pid = pid + 1
         continue
     pid = pid + 1
-    interval_minute=interval/60
-    interval_second=interval%60
+    interval_minute=int(interval/60)
+    interval_second=int(interval%60)
     for i in range(interval_minute):
         a.sendheartbeat()
         time.sleep(60)
